refactor(base): flatten node maintenance commands to top level
Remove the 'base node <cmd>' subcommand group in favour of flat commands:
base db, base init, base init-state, base dump-genesis, base stage,
base p2p, base config, base prune, base re-execute, base proofs,
base snapshot-manifest, base download.

Each command constructs its own CliRunner. The --chain flag on these
reth-derived commands comes from their embedded BaseChainSpecParser,
not from BaseCli's global --chain (clap skips global propagation when
the child already owns the same arg id).

Remove the rejects_legacy_node_rpc_path test — the 'node' subgroup no
longer exists.
