+++ bin/base/src/commands/mod.rs
@@ -3,7 +3,6 @@
 mod bootnode;
 mod command;
 pub(crate) use command::BaseCommand;
-mod node;
 mod rpc;
 mod sequencer;
 mod update;
